Update figwheel-main watch directory
This suppresses the error about src/minimal_example/cljs not being on the classpath, though at the cost of having figwheel-main reload the page on clj file edits as well. Setting :source-paths to ["src/minimal_example/clj" "src/minimal_example/cljs"] as alluded to by the figwheel-main error message causes lein to be unable to find the minimal_example.clj.core, so perhaps an incompatibility between how the two want to manage namespaces?